requirements.txt
Django==2.2.10
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv==0.17.1
pytest-django==4.4.0
pytest==6.2.4
